Remove base plugin class
